where2b_backend/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0
drf-yasg==1.20.0 
psycopg2-binary>=2.8
